Updated code to use standard (According to Gemini)

+++ src/main/java/com/seevrantillan/safespace/repository/ReactionRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.seevrantillan.safespace.entity.ReactionEntity;
 
 @Repository
-public interface ReactionRepository extends JpaRepository<ReactionEntity, Integer> {
+public interface ReactionRepository extends JpaRepository<ReactionEntity, Long> {
 }
